requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-Cors==4.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.4
numpy==1.26.3
pandas==2.2.0
psycopg-binary==3.1.18
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
scipy==1.12.0
six==1.16.0
tzdata==2023.4
waitress==3.0.0
Werkzeug==3.0.1
xgboost==2.0.3
zipp==3.17.0
